factory: view코드 수정
