+++ Plantory/Plantory/Views/Profile/ProfileViews/TempViews/TempStorageView.swift
@@ -1,20 +1,20 @@
 import SwiftUI
 
-struct TrashView: View {
-    private let container: DIContainer
+struct TempStorageView: View {
     @Environment(\.dismiss) private var dismiss
-    @StateObject private var viewModel: WasteViewModel
+    private let container: DIContainer
+    
+    @StateObject private var viewModel: TempViewModel
     
     init(container: DIContainer) {
             self.container = container
-            _viewModel = StateObject(wrappedValue: WasteViewModel(container: container))
+            _viewModel = StateObject(wrappedValue: TempViewModel(container: container))
         }
-
+    
     @State private var isNewSorting = true
     @State private var isEditing = false
     @State private var checkedItems = Set<Int>()
-    @State private var showDeletePopUp = false
-    @State private var showRestorePopUp = false
+    @State private var showPopUp = false
 
     var body: some View {
         ZStack {
@@ -23,29 +23,30 @@ struct TrashView: View {
 
                 content
 
-                TrashFootView(
+                TempFootView(
                     isEditing: $isEditing,
                     isEmpty: checkedItems.isEmpty,
-                    onRestore: { showRestorePopUp = true },
-                    onDelete: { showDeletePopUp = true }
+                    onDelete: { showPopUp = true }
                 )
             }
             .padding(.horizontal)
             .onChange(of: isNewSorting) { _, newValue in
-                viewModel.fetchWaste(sort: newValue ? .latest : .oldest)
+                viewModel.fetchTemp(sort: newValue ? .latest : .oldest)
             }
             .customNavigation(
-                title: "휴지통",
+                title: "임시보관함",
                 leading: navigationLeading,
                 trailing: navigationTrailing
             )
             .navigationBarBackButtonHidden(true)
 
-            if showDeletePopUp { deleteConfirmationPopUp }
-            if showRestorePopUp { restoreConfirmationPopUp }
+            if showPopUp {
+                deleteConfirmationPopUp
+            }
         }
     }
 
+    // MARK: - 콘텐츠 뷰
     @ViewBuilder
     private var content: some View {
         if viewModel.isLoading {
@@ -53,14 +54,14 @@ struct TrashView: View {
         } else if let error = viewModel.errorMessage {
             Text(error).foregroundColor(.red)
         } else if sortedCells.isEmpty {
-            NothingView(mainText: "휴지통이 비어있습니다", subText: "삭제된 일기가 없습니다.")
+            NothingView(mainText: "보관한 일기가 없어요", subText: "작성 중인 일기를 보관함에 저장해 놓을 수 있어요!")
                 .frame(maxWidth: .infinity, maxHeight: .infinity)
         } else {
             diaryList
         }
     }
 
-    private var sortedCells: [WasteViewModel.DiaryCellViewModel] {
+    private var sortedCells: [TempViewModel.DiaryCellViewModel] {
         viewModel.cellViewModels.sorted(by: isNewSorting ? { $0.dateText > $1.dateText } : { $0.dateText < $1.dateText })
     }
 
@@ -75,8 +76,11 @@ struct TrashView: View {
                         isChecked: Binding(
                             get: { checkedItems.contains(cell.id) },
                             set: { isChecked in
-                                if isChecked { checkedItems.insert(cell.id) }
-                                else { checkedItems.remove(cell.id) }
+                                if isChecked {
+                                    checkedItems.insert(cell.id)
+                                } else {
+                                    checkedItems.remove(cell.id)
+                                }
                             }
                         ),
                         onNavigate: {}
@@ -88,12 +92,13 @@ struct TrashView: View {
         }
     }
 
+    // MARK: - 네비게이션 바 구성요소
     private var navigationLeading: some View {
         Group {
             if isEditing {
                 Button(action: toggleAllSelection) {
                     Text(checkedItems.count == sortedCells.count ? "전체 선택 해제" : "전체 선택")
-                        .font(.pretendardRegular(14)).foregroundStyle(.green07)
+                        .font(.pretendardRegular(16)).foregroundStyle(.green07)
                 }
             } else {
                 Button(action: dismiss.callAsFunction) {
@@ -106,32 +111,23 @@ struct TrashView: View {
     private var navigationTrailing: some View {
         Button(action: { isEditing.toggle() }) {
             Text(isEditing ? "취소" : "편집")
-                .font(.pretendardRegular(14)).foregroundStyle(.green07)
+                .font(.pretendardRegular(16)).foregroundStyle(.green07)
         }
     }
 
+    // MARK: - 삭제 확인 팝업
     private var deleteConfirmationPopUp: some View {
         PopUp(
-            title: "일기를 삭제하시겠습니까?",
-            message: "일기 삭제 시, 해당 일기는 영구 삭제됩니다.",
+            title: "보관한 일기를 삭제하시겠습니까?",
+            message: "일기 삭제 시, 해당 일기는 휴지통으로 이동합니다.",
             confirmTitle: "삭제하기",
             cancelTitle: "취소",
             onConfirm: performDeletion,
-            onCancel: { showDeletePopUp = false }
+            onCancel: { showPopUp = false }
         )
     }
 
-    private var restoreConfirmationPopUp: some View {
-        PopUp(
-            title: "해당 일기를 복원하시겠습니까?",
-            message: "일기 복원 시, 해당 일기는 유지됩니다.",
-            confirmTitle: "복원하기",
-            cancelTitle: "취소",
-            onConfirm: performRestore,
-            onCancel: { showRestorePopUp = false }
-        )
-    }
-
+    // MARK: - 액션
     private func toggleAllSelection() {
         if checkedItems.count == sortedCells.count {
             checkedItems.removeAll()
@@ -141,49 +137,34 @@ struct TrashView: View {
     }
 
     private func performDeletion() {
-        viewModel.deleteForever(ids: Array(checkedItems))
-        // 실제로 삭제할 때는 fetch()로 수정된 리스트 불러오기 !!
+        viewModel.moveToTrash(ids: Array(checkedItems))
         checkedItems.removeAll()
         isEditing = false
-        showDeletePopUp = false
-    }
-
-    private func performRestore() {
-        // 복원 API 호출
-        viewModel.restoreWaste(ids: Array(checkedItems))
-        checkedItems.removeAll()
-        isEditing = false
-        showRestorePopUp = false
+        showPopUp = false
     }
 }
 
-/// 하단 복원/삭제 버튼 뷰
-struct TrashFootView: View {
+
+struct TempFootView: View {
     @Binding var isEditing: Bool
     let isEmpty: Bool
-    let onRestore: () -> Void
-    let onDelete: () -> Void
-
+    var onDelete: () -> Void   // 삭제 액션 클로저
+    
     var body: some View {
         VStack {
             if isEditing {
                 HStack {
-                    MainSmallButton(
-                        text: "복원하기",
-                        isDisabled: isEmpty,
-                        action: onRestore
-                    )
                     Spacer()
+
                     MainSmallButton(
                         text: "삭제",
                         isDisabled: isEmpty,
                         action: onDelete
                     )
                 }
-
             } else {
                 HStack {
-                    Text("휴지통에 있는 항목은 이동된 날짜로부터 30일 뒤 영구삭제 됩니다.")
+                    Text("보관함에 있는 항목은 이동된 날짜로부터 30일 뒤 휴지통으로 이동합니다.")
                         .font(.pretendardLight(12))
                         .foregroundColor(.gray08)
                         .padding(.vertical, 11)
@@ -194,6 +175,7 @@ struct TrashFootView: View {
     }
 }
 
+
 #Preview {
-    NavigationStack { TrashView(container: .init()) }
+    NavigationStack { TempStorageView(container: .init()) }
 }
